+++ tests/test_basic_functionality.py
@@ -115,3 +115,33 @@ def test_m_u_chart(api_info, fake_1000_factory, fake_1000_settings):
     linker = Linker(df, fake_1000_settings, db_api)
 
     linker.visualisations.m_u_parameters_chart()
+
+
+def test_comparison_viewer_dashboard(
+    api_info, fake_1000_factory, fake_1000_settings, tmp_path
+):
+    db_api = api_info["db_api"]
+    df = fake_1000_factory(api_info["version"])
+    fake_1000_settings.retain_intermediate_calculation_columns = True
+    linker = Linker(df, fake_1000_settings, db_api)
+
+    df_predict = linker.inference.predict()
+    linker.visualisations.comparison_viewer_dashboard(df_predict, tmp_path / "cvd.html")
+
+
+def test_cluster_studio_dashboard(
+    api_info, fake_1000_factory, fake_1000_settings, tmp_path
+):
+    db_api = api_info["db_api"]
+    df = fake_1000_factory(api_info["version"])
+    fake_1000_settings.retain_intermediate_calculation_columns = True
+    linker = Linker(df, fake_1000_settings, db_api)
+
+    df_predict = linker.inference.predict()
+    df_clustered = linker.clustering.cluster_pairwise_predictions_at_threshold(
+        df_predict,
+        threshold_match_probability=0.8,
+    )
+    linker.visualisations.cluster_studio_dashboard(
+        df_predict, df_clustered, tmp_path / "csd.html"
+    )